BUG: Added some fixes to CMakeLists.txt files
This commit fixes a few issues that prevented me from compiling
correctly on the University of Manchester CSF cluster. One issue was
around correctly linking HDF5 libraries, where it seems that the
HDF5_CXX_INCLUDE_DIR variable has been deprecated in favour of the
HDF5_CXX_INCLUDE_DIRS variable. The second issue relates to pjemris,
which requires the pthread library to be linked. This is a dependency of
the HDF5 library available through apt, so is normally linked
automatically. However when installing HDF5 from source (as I did not
have admin privileges) the same is not true. Instead I have added a
find_package(Threads) which will add the pthreads library independently.